Fix virtual operation for Steem 0.22.1
