+++ core/src/main/java/eu/bittrade/libs/steemj/protocol/operations/virtual/CurationRewardOperation.java
@@ -23,12 +23,14 @@ import org.apache.commons.lang3.builder.ToStringBuilder;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import eu.bittrade.libs.steemj.base.models.Permlink;
 import eu.bittrade.libs.steemj.enums.PrivateKeyType;
 import eu.bittrade.libs.steemj.enums.ValidationType;
 import eu.bittrade.libs.steemj.exceptions.SteemInvalidTransactionException;
 import eu.bittrade.libs.steemj.interfaces.SignatureObject;
+import eu.bittrade.libs.steemj.protocol.AccountName;
+import eu.bittrade.libs.steemj.protocol.Asset;
 import eu.bittrade.libs.steemj.protocol.operations.Operation;
-import eu.bittrade.libs.steemj.protocol.operations.virtual.value.CurationRewardValue;
 
 /**
  * This class represents a "curation_reward_operation" object.
@@ -39,8 +41,14 @@ import eu.bittrade.libs.steemj.protocol.operations.virtual.value.CurationRewardV
  * @author <a href="http://steemit.com/@dez1337">dez1337</a>
  */
 public class CurationRewardOperation extends Operation {
-	@JsonProperty("value")
-	private CurationRewardValue curationRewardValue;
+    @JsonProperty("curator")
+    private AccountName curator;
+    @JsonProperty("reward")
+    private Asset reward;
+    @JsonProperty("comment_author")
+    private AccountName commentAuthor;
+    @JsonProperty("comment_permlink")
+    private Permlink commentPermlink;
 
     /**
      * This operation is a virtual one and can only be created by the blockchain
@@ -51,14 +59,43 @@ public class CurationRewardOperation extends Operation {
     }
 
     /**
-     * Gets all the values 
+     * Get the person that receives the reward.
      * 
-     * @return the value for type curation_reward_operation
+     * @return The person that receives the reward.
      */
-	public CurationRewardValue getCurationRewardValue() {
-		return curationRewardValue;
-	}
-	
+    public AccountName getCurator() {
+        return curator;
+    }
+
+    /**
+     * Get the amount and the currency the curator receives.
+     * 
+     * @return The reward.
+     */
+    public Asset getReward() {
+        return reward;
+    }
+
+    /**
+     * Get the author of the post or comment that this curation reward is for.
+     * 
+     * @return The author of the post or comment.
+     */
+    public AccountName getCommentAuthor() {
+        return commentAuthor;
+    }
+
+    /**
+     * Get the permanent link of the post or comment that this curation reward
+     * is for.
+     * 
+     * @return The permanent link of the post or comment that this curation
+     *         reward is for.
+     */
+    public Permlink getCommentPermlink() {
+        return commentPermlink;
+    }
+
     @Override
     public byte[] toByteArray() throws SteemInvalidTransactionException {
         // The byte representation is not needed for virtual operations as we
