+++ core/src/main/java/eu/bittrade/libs/steemj/protocol/operations/virtual/FillConvertRequestOperation.java
@@ -27,8 +27,9 @@ import eu.bittrade.libs.steemj.enums.PrivateKeyType;
 import eu.bittrade.libs.steemj.enums.ValidationType;
 import eu.bittrade.libs.steemj.exceptions.SteemInvalidTransactionException;
 import eu.bittrade.libs.steemj.interfaces.SignatureObject;
+import eu.bittrade.libs.steemj.protocol.AccountName;
+import eu.bittrade.libs.steemj.protocol.Asset;
 import eu.bittrade.libs.steemj.protocol.operations.Operation;
-import eu.bittrade.libs.steemj.protocol.operations.virtual.value.FillConvertRequestOperationValue;
 
 /**
  * This class represents a Steem "fill_convert_request_operation" object.
@@ -39,8 +40,15 @@ import eu.bittrade.libs.steemj.protocol.operations.virtual.value.FillConvertRequ
  * @author <a href="http://steemit.com/@dez1337">dez1337</a>
  */
 public class FillConvertRequestOperation extends Operation {
-    @JsonProperty("value")
-    private FillConvertRequestOperationValue value;
+    @JsonProperty("owner")
+    private AccountName owner;
+    // Original type is uint32_t so we have to use long here.
+    @JsonProperty("requestid")
+    private long requestId;
+    @JsonProperty("amount_in")
+    private Asset amountIn;
+    @JsonProperty("amount_out")
+    private Asset amountOut;
 
     /**
      * This operation is a virtual one and can only be created by the blockchain
@@ -51,15 +59,43 @@ public class FillConvertRequestOperation extends Operation {
     }
 
     /**
-     * Gets all the values 
+     * Get the owner of this conversion request.
      * 
-     * @return the value for type fill_convert_request_operation
+     * @return The owner as an AccountName instance.
      */
-    public FillConvertRequestOperationValue getValue() {
-		return value;
-	}
+    public AccountName getOwner() {
+        return owner;
+    }
 
-	@Override
+    /**
+     * Get the id of this request.
+     * 
+     * @return The id of this request.
+     */
+    public long getRequestId() {
+        return requestId;
+    }
+
+    /**
+     * Get the amount and the type of the currency that has been converted
+     * within this operation.
+     * 
+     * @return The source asset.
+     */
+    public Asset getAmountIn() {
+        return amountIn;
+    }
+
+    /**
+     * Get the amount and the type of the target currency.
+     * 
+     * @return The target asset.
+     */
+    public Asset getAmountOut() {
+        return amountOut;
+    }
+
+    @Override
     public byte[] toByteArray() throws SteemInvalidTransactionException {
         // The byte representation is not needed for virtual operations as we
         // can't broadcast them.
