+++ core/src/main/java/eu/bittrade/libs/steemj/protocol/operations/virtual/ProducerRewardOperation.java
@@ -28,8 +28,9 @@ import eu.bittrade.libs.steemj.enums.PrivateKeyType;
 import eu.bittrade.libs.steemj.enums.ValidationType;
 import eu.bittrade.libs.steemj.exceptions.SteemInvalidTransactionException;
 import eu.bittrade.libs.steemj.interfaces.SignatureObject;
+import eu.bittrade.libs.steemj.protocol.AccountName;
+import eu.bittrade.libs.steemj.protocol.Asset;
 import eu.bittrade.libs.steemj.protocol.operations.Operation;
-import eu.bittrade.libs.steemj.protocol.operations.virtual.value.ProducerRewardOperationValue;
 
 /**
  * This class represents the Steem "producer_reward_operation" object.
@@ -40,8 +41,10 @@ import eu.bittrade.libs.steemj.protocol.operations.virtual.value.ProducerRewardO
  */
 @JsonTypeName("producer_reward_operation")
 public class ProducerRewardOperation extends Operation {
-	@JsonProperty("value")
-	private ProducerRewardOperationValue value;
+    @JsonProperty("producer")
+    private AccountName producer;
+    @JsonProperty("vesting_shares")
+    private Asset vestingShares;
 
     /**
      * This operation is a virtual one and can only be created by the blockchain
@@ -52,16 +55,24 @@ public class ProducerRewardOperation extends Operation {
     }
 
     /**
-     * Gets all the values 
+     * Get the block producer.
      * 
-     * @return the value for type producer_reward_operation
+     * @return The block producer.
      */
-    public ProducerRewardOperationValue getValue() {
-		return value;
-	}
+    public AccountName getProducer() {
+        return producer;
+    }
 
+    /**
+     * Get the amount of VESTS the <code>producer</code> got.
+     * 
+     * @return The vesting shares paid to the <code>producer</code>.
+     */
+    public Asset getVestingShares() {
+        return vestingShares;
+    }
 
-	@Override
+    @Override
     public byte[] toByteArray() throws SteemInvalidTransactionException {
         // The byte representation is not needed for virtual operations as we
         // can't broadcast them.
