+++ core/src/main/java/eu/bittrade/libs/steemj/protocol/operations/virtual/ReturnVestingDelegationOperation.java
@@ -27,8 +27,9 @@ import eu.bittrade.libs.steemj.enums.PrivateKeyType;
 import eu.bittrade.libs.steemj.enums.ValidationType;
 import eu.bittrade.libs.steemj.exceptions.SteemInvalidTransactionException;
 import eu.bittrade.libs.steemj.interfaces.SignatureObject;
+import eu.bittrade.libs.steemj.protocol.AccountName;
+import eu.bittrade.libs.steemj.protocol.Asset;
 import eu.bittrade.libs.steemj.protocol.operations.Operation;
-import eu.bittrade.libs.steemj.protocol.operations.virtual.value.ReturnVestingDelegationOperationValue;
 
 /**
  * This class represents the Steem "return_vesting_delegation_operation" object.
@@ -36,8 +37,9 @@ import eu.bittrade.libs.steemj.protocol.operations.virtual.value.ReturnVestingDe
  * @author <a href="http://steemit.com/@dez1337">dez1337</a>
  */
 public class ReturnVestingDelegationOperation extends Operation {
-	@JsonProperty("value")
-    private ReturnVestingDelegationOperationValue value;
+    private AccountName account;
+    @JsonProperty("vesting_shares")
+    private Asset vestingShares;
 
     /**
      * This operation is a virtual one and can only be created by the blockchain
@@ -48,17 +50,20 @@ public class ReturnVestingDelegationOperation extends Operation {
     }
 
     /**
-     * Gets all the values 
-     * 
-     * @return the value for type return_vesting_delegation_operation
+     * @return The account which receives the VESTS.
      */
-    public ReturnVestingDelegationOperationValue getValue() {
-		return value;
-	}
+    public AccountName getAccount() {
+        return account;
+    }
 
+    /**
+     * @return Return the amount of VESTS the account receives.
+     */
+    public Asset getVestingShares() {
+        return vestingShares;
+    }
 
-
-	@Override
+    @Override
     public byte[] toByteArray() throws SteemInvalidTransactionException {
         // The byte representation is not needed for virtual operations as we
         // can't broadcast them.
